advance to the first position for a document
Since we can index phrases, a term might occur multiple times in a
document. We need to make sure we're looking at the first one when
advancing.  We might miss a hit otherwise during a phrase search
if a position we skipped over was in fact one we wanted to match.